refactor: API 5-8 DTO 최소화 및 평점 시스템 개선
- 평점 시스템을 Double(0.5-5.0)에서 Int(1-10)로 변경
- API 5-8에 필요한 최소한의 DTO만 생성
- 불필요한 Response DTO 삭제 (제목 변경 등은 응답 불필요)
- SendMessageResponse는 세션 제목 포함 (첫 메시지 시에만)
- RateSessionRequest 유효성 검사 강화

+++ src/main/kotlin/com/aicounseling/app/domain/counselor/entity/CounselorRating.kt
@@ -26,8 +26,9 @@ class CounselorRating(
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "session_id", nullable = false)
     val session: ChatSession,
+    // 1~10 (별 0.5개 = 1, 별 5개 = 10)
     @Column(nullable = false)
-    val rating: Double,
+    val rating: Int,
     @Column(columnDefinition = "TEXT")
     val review: String? = null,
 ) : BaseEntity()
